# Optimize rendering