--- a/dependencies.gradle
+++ b/dependencies.gradle
@@ -1,5 +1,5 @@
 // Add your dependencies here
 
 dependencies {
-    implementation('com.github.GTNewHorizons:GTNHLib:0.0.10:dev')
+    implementation('com.github.GTNewHorizons:GTNHLib:0.0.12:dev')
 }