--- a/repositories.gradle
+++ b/repositories.gradle
@@ -1,9 +1,4 @@
 // Add any additional repositories for your dependencies here
 
 repositories {
-    maven {
-        name "GTNH Maven"
-        url "http://jenkins.usrv.eu:8081/nexus/content/groups/public/"
-        allowInsecureProtocol true
-    }
 }