On a superflat world in the malisisdoors environement with ~100 doors with these changes I go from 100 to 220 FPS, and allocation rate reduces from ~1.2GB/s to 800MB/s. Any more changes would require more significant refactoring of the code, but this should make door rendering less of a bottleneck.

I change a protected field type from a List to an array, but it looks like none of the malisis mods use it or even subclass RenderParameters, so it shouldn't cause ABI breaks.